events, guests and einvite page updated
