+++ app/components/AddEventModal/index.js
@@ -18,6 +18,9 @@ function AddEventModal({
   eventDetails,
   index,
   updateEvent,
+  isUpdate,
+  setIsUpdate,
+  events,
 }) {
   useInjectReducer({ key: "eventsPage", reducer });
   useInjectSaga({ key: "eventsPage", saga });
@@ -35,10 +38,20 @@ function AddEventModal({
   });
   const onInputChange = (e) => {
     const { name, value } = e.target;
-    setInput((prev) => ({
-      ...prev,
-      [name]: value,
-    }));
+    setInput((prev) => {
+      if(name === "category" && value !== "other")
+      {
+        return {
+          ...prev,
+          [name]: value,
+          "customEvent": "",
+        }; 
+      }
+      return {
+        ...prev,
+        [name]: value,
+      };
+    });
   };
 
   const handleSubmit = async (event) => {
@@ -51,6 +64,7 @@ function AddEventModal({
         input.time,
         input.eventVenue
       );
+      setIsOpen(!isOpen);
     } else {
       await updateEvent({
         id: eventDetails._id,
@@ -60,7 +74,7 @@ function AddEventModal({
         time: input.time,
         venue: input.eventVenue,
       });
-      window.location.reload();
+      setIsUpdate(new Array(events.length).fill(false));
     }
   };
 
@@ -68,9 +82,7 @@ function AddEventModal({
     if (role === "add") {
       setIsOpen(!isOpen);
     } else {
-      setIsOpen((prevState) =>
-        prevState.map((item, idx) => (idx === index ? !item : item))
-      );
+      setIsUpdate(new Array(events.length).fill(false));
     }
   };
 
@@ -78,16 +90,15 @@ function AddEventModal({
     <div className="overflow-y-hidden overflow-x-hidden fixed top-1/2 left-1/2 z-40 w-4/5 lg:w-1/3 -translate-x-1/2 -translate-y-1/2 bg-white flex flex-col py-4 pl-8 rounded-lg">
       <AiOutlineCloseCircle
         onClick={() => handleClick(index)}
-        className="absolute top-0 right-0 m-2"
+        className="absolute top-0 right-0 m-2 cursor-pointer"
       />
       <h3 className="mt-3 mb-2 md:mb-4 text-xl font-medium text-gray-900">
         {role === "add" ? "Add" : "Update"} Event
       </h3>
       <form
-        class="space-y-3 md:space-y-6"
+        className="space-y-3 md:space-y-6"
         onSubmit={(event) => {
           handleSubmit(event);
-          setIsOpen(!isOpen);
         }}
       >
         <div>
@@ -105,15 +116,17 @@ function AddEventModal({
             className="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-11/12 p-1.5 py-2 md:p-2.5"
           >
             <option>Category</option>
-            {options.map((option) => (
-              <option value={option.value}>{option.label}</option>
+            {options.map((option, index) => (
+              <option value={option.value} key={index}>
+                {option.label}
+              </option>
             ))}
           </select>
         </div>
         {input.category === "Other" && (
           <div>
             <label
-              for="customEvent"
+              htmlFor="customEvent"
               className="block mb-1 md:mb-2 text-sm font-medium text-gray-900"
             >
               Event Name
@@ -131,7 +144,7 @@ function AddEventModal({
         )}
         <div>
           <label
-            for="date"
+            htmlFor="date"
             className="block mb-1 md:mb-2 text-sm font-medium text-gray-900"
           >
             Event Date
@@ -149,7 +162,7 @@ function AddEventModal({
         </div>
         <div>
           <label
-            for="time"
+            htmlFor="time"
             className="block mb-1 md:mb-2 text-sm font-medium text-gray-900"
           >
             Event Time
@@ -166,7 +179,7 @@ function AddEventModal({
         </div>
         <div>
           <label
-            for="eventVenue"
+            htmlFor="eventVenue"
             className="block mb-1 md:mb-2 text-sm font-medium text-gray-900"
           >
             Event Venue
