--- app/containers/MobileEventsPage/Loadable.js
@@ -1,9 +1,0 @@
-/**
- *
- * Asynchronously loads the component for MobileEventsPage
- *
- */
-
-import loadable from "utils/loadable";
-
-export default loadable(() => import("./index"));
