--- app/containers/MobileEventsPage/actions.js
@@ -1,13 +1,0 @@
-/*
- *
- * MobileEventsPage actions
- *
- */
-
-import { DEFAULT_ACTION } from "./constants";
-
-export function defaultAction() {
-  return {
-    type: DEFAULT_ACTION
-  };
-}
