--- app/containers/MobileEventsPage/constants.js
@@ -1,7 +1,0 @@
-/*
- *
- * MobileEventsPage constants
- *
- */
-
-export const DEFAULT_ACTION = "app/MobileEventsPage/DEFAULT_ACTION";
