--- app/containers/MobileEventsPage/messages.js
@@ -1,16 +1,0 @@
-/*
- * MobileEventsPage Messages
- *
- * This contains all the text for the MobileEventsPage container.
- */
-
-import { defineMessages } from "react-intl";
-
-export const scope = "app.containers.MobileEventsPage";
-
-export default defineMessages({
-  header: {
-    id: `${scope}.header`,
-    defaultMessage: "This is the MobileEventsPage container!"
-  }
-});
