--- app/containers/MobileEventsPage/reducer.js
@@ -1,20 +1,0 @@
-/*
- *
- * MobileEventsPage reducer
- *
- */
-import produce from "immer";
-import { DEFAULT_ACTION } from "./constants";
-
-export const initialState = {};
-
-/* eslint-disable default-case, no-param-reassign */
-const mobileEventsPageReducer = (state = initialState, action) =>
-  produce(state, (/* draft */) => {
-    switch (action.type) {
-      case DEFAULT_ACTION:
-        break;
-    }
-  });
-
-export default mobileEventsPageReducer;
