--- app/containers/MobileEventsPage/saga.js
@@ -1,6 +1,0 @@
-// import { take, call, put, select } from 'redux-saga/effects';
-
-// Individual exports for testing
-export default function* mobileEventsPageSaga() {
-  // See example in containers/HomePage/saga.js
-}
