--- app/containers/MobileEventsPage/tests/actions.test.js
@@ -1,13 +1,0 @@
-import { defaultAction } from "../actions";
-import { DEFAULT_ACTION } from "../constants";
-
-describe("MobileEventsPage actions", () => {
-  describe("Default Action", () => {
-    it("has a type of DEFAULT_ACTION", () => {
-      const expected = {
-        type: DEFAULT_ACTION
-      };
-      expect(defaultAction()).toEqual(expected);
-    });
-  });
-});
