--- app/containers/MobileEventsPage/tests/selectors.test.js
@@ -1,7 +1,0 @@
-// import { selectMobileEventsPageDomain } from '../selectors';
-
-describe("selectMobileEventsPageDomain", () => {
-  it("Expect to have unit tests specified", () => {
-    expect(true).toEqual(false);
-  });
-});
