--- app/containers/MobileGuestsPage/actions.js
@@ -1,13 +1,0 @@
-/*
- *
- * MobileGuestsPage actions
- *
- */
-
-import { DEFAULT_ACTION } from "./constants";
-
-export function defaultAction() {
-  return {
-    type: DEFAULT_ACTION
-  };
-}
