--- app/containers/MobileGuestsPage/constants.js
@@ -1,7 +1,0 @@
-/*
- *
- * MobileGuestsPage constants
- *
- */
-
-export const DEFAULT_ACTION = "app/MobileGuestsPage/DEFAULT_ACTION";
