--- app/containers/MobileGuestsPage/messages.js
@@ -1,16 +1,0 @@
-/*
- * MobileGuestsPage Messages
- *
- * This contains all the text for the MobileGuestsPage container.
- */
-
-import { defineMessages } from "react-intl";
-
-export const scope = "app.containers.MobileGuestsPage";
-
-export default defineMessages({
-  header: {
-    id: `${scope}.header`,
-    defaultMessage: "This is the MobileGuestsPage container!"
-  }
-});
