--- app/containers/MobileGuestsPage/tests/saga.test.js
@@ -1,15 +1,0 @@
-/**
- * Test sagas
- */
-
-/* eslint-disable redux-saga/yield-effects */
-// import { take, call, put, select } from 'redux-saga/effects';
-// import mobileGuestsPageSaga from '../saga';
-
-// const generator = mobileGuestsPageSaga();
-
-describe("mobileGuestsPageSaga Saga", () => {
-  it("Expect to have unit tests specified", () => {
-    expect(true).toEqual(false);
-  });
-});
