--- app/containers/MobileGuestsPage/tests/selectors.test.js
@@ -1,7 +1,0 @@
-// import { selectMobileGuestsPageDomain } from '../selectors';
-
-describe("selectMobileGuestsPageDomain", () => {
-  it("Expect to have unit tests specified", () => {
-    expect(true).toEqual(false);
-  });
-});
