+++ app/utils/searchByName.js
@@ -1,0 +1,4 @@
+export default function searchByName(guests, text) {
+  text = text.toLowerCase();
+  return guests.filter((guest) => guest.name.toLowerCase().includes(text));
+}
