+++ package.json
@@ -98,6 +98,7 @@
     "sanitize.css": "^8.0.0",
     "slick-carousel": "^1.8.1",
     "styled-components": "4.2.0",
+    "use-debounce": "^8.0.4",
     "use-onclickoutside": "^0.4.1"
   },
   "devDependencies": {
